RavenDB-5055 - fix typo collecet -> collect - in BuildExpression, check if InvocationExpression.Expression is of type MemberAccessExpressionSyntax or IdentifierNameSyntax - change CaptureSelectNewFieldNamesVisitor to CaptureSelectExpressionsAndNewFieldNamesVisitor and use the old CaptureSelectNewFieldNamesVisitor from Indexes.Static.Roslyn.Rewriters when possible
